Clean up specs config files after fixing Travis chromedriver error
